+++ exporters/otlp/otlptrace/internal/tracetransform/span_test.go
@@ -298,6 +298,17 @@ func TestSpanData(t *testing.T) {
 
 	assert.Equal(t, got[0].GetResource(), Resource(spanData.Resource))
 	assert.Equal(t, got[0].SchemaUrl, spanData.Resource.SchemaURL())
+
+	ilSpans := got[0].GetInstrumentationLibrarySpans() //nolint:staticcheck
+	require.Len(t, ilSpans, 1)
+	assert.Equal(t, ilSpans[0].SchemaUrl, spanData.InstrumentationLibrary.SchemaURL)
+	assert.Equal(t, ilSpans[0].GetInstrumentationLibrary(), InstrumentationLibrary(spanData.InstrumentationLibrary))
+	require.Len(t, ilSpans[0].Spans, 1)
+	actualIlSpan := ilSpans[0].Spans[0]
+	if diff := cmp.Diff(expectedSpan, actualIlSpan, cmp.Comparer(proto.Equal)); diff != "" {
+		t.Fatalf("transformed span differs %v\n", diff)
+	}
+
 	scopeSpans := got[0].GetScopeSpans()
 	require.Len(t, scopeSpans, 1)
 	assert.Equal(t, scopeSpans[0].SchemaUrl, spanData.InstrumentationLibrary.SchemaURL)
@@ -315,6 +326,12 @@ func TestRootSpanData(t *testing.T) {
 	sd := Spans(tracetest.SpanStubs{{}}.Snapshots())
 	require.Len(t, sd, 1)
 	rs := sd[0]
+
+	ilSpans := rs.GetInstrumentationLibrarySpans() // nolint:staticcheck
+	require.Len(t, ilSpans, 1)
+	ilGot := rs.GetInstrumentationLibrarySpans()[0].GetSpans()[0].GetParentSpanId() // nolint:staticcheck
+	assert.Nil(t, ilGot, "incorrect transform of root parent span ID")
+
 	scopeSpans := rs.GetScopeSpans()
 	require.Len(t, scopeSpans, 1)
 	got := scopeSpans[0].GetSpans()[0].GetParentSpanId()